Fix parameters for proper remarketing tracking
According to the adwords docs the `lat` parameter means `limited ad tracking` and should be set to `1` if the user disabled ad tracking in his iOS settings. In your implementation it is used exactly the other way around. If the user allows ad tracking it was set to `1` whereas it should have been set to `0`.

Also according to the docs the parameter `usage_tracking_enabled` is required, so I added it.

https://developers.google.com/app-conversion-tracking/ios/remarketing-server